+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/resource-package/compare/0.1.0...develop)
+## [Unreleased](https://github.com/ixocreate/resource-package/compare/0.2.0...develop)
+
+## [v0.2.0 (2019-05-06)](https://github.com/ixocreate/resource-package/compare/0.1.0...0.2.0)
+### Added
+- Compatibility with Application v0.2
+### Changed
+- Upgrade to Database Package v0.2
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/resource-package/compare/master...0.1.0)
-
 ### Changed
 - Consolidate Package